去掉“最常使用列表”，因为最常使用一难算而意义不大

+++ DataMag/DataMag.cpp
@@ -134,7 +134,7 @@ void CSetting::SetRecentFile(CString file)
 
 	recentFileList.insert(recentFileList.begin(), file);
 
-	for (int i = 0; i < recentFileList.size(); i++) {
+	for (size_t i = 0; i < recentFileList.size(); i++) {
 
 		CString recentFileIndex;
 		recentFileIndex.Format(_T("File%d"), i);
